Extract transaction controller
