+++ slides/decks/02/index.html
@@ -152,6 +152,28 @@
 					</section>
 					<section>
 						<h2>Connections</h2>
+						<pre data-id="code-animation"><code class='typescript' data-trim data-line-numbers="|1|3-9|11-17|19|"><script type="text/template">
+							import { Schema, model } from 'mongoose'
+
+							export interface User {
+								name: {
+									first: string,
+									last: string,
+								},
+								email: string
+							}
+
+							const schema = new Schema<User>({
+								name: { 
+									first: { type: String, required: true },
+									last: { type: String, required: true }
+								},
+								email: { type: String, required: false },
+							})
+
+							export const TransactionModel = model<User>('User', schema)
+						</script></code></pre>
+						<small>examples/lesson-02/hello-mongoose/src/user.ts</small>
 					</section>
 					<section>
 						<h2>Schemas</h2>
